@@ -12,6 +12,7 @@
 #include "Complex.hpp"
 #include "Matrix.hpp"
 #include "LSMSMode.hpp"
+#include "LSMSAlgorithms.hpp"
 
 #include "SingleSite/AtomData.hpp"
 
@@ -64,6 +65,7 @@ public:
   char systemid[80];
   char title[80];
   LSMSMode lsmsMode;
+  signed char lsmsAlgorithms[LSMSAlgorithms::numAlgorithms];
   int rank;
   char potential_file_in[128];
   char potential_file_out[128];
@@ -118,7 +120,7 @@ public:
   Real clight;
   int maxlmax;
   LSMSGlobals global;
-  AngularMomentumIndices angularMomentumIndices;
+  Matrix<Real> dl_factor;
   EnergyContourParameters energyContour;
 
 // no. of Gaussian points for volume integration
@@ -133,6 +135,8 @@ public:
   Real volumeNorm;             // Volume renormalization factor
   Real volumeInterstitial;     // Total interstitial volume
   Real u0;                     // Contribution of the Muffin-tin zero potential to the Coulomb energy
+  Real u0MT;                   // Contribution of the Muffin-tin zero potential to the Coulomb energy just from MT
+  Real vmt;                    // Muffin-tin zero
   Real totalEnergy;            // Total energy
   //Real pressure;               // Pressure
 
