@@ -8,6 +8,9 @@
 #include "Matrix.hpp"
 #include <vector>
 
+#include "Misc/Coeficients.hpp"
+#include "Misc/Indices.hpp"
+
 #include "Accelerator/DeviceStorage.hpp"
 #include <hip/hip_runtime.h>
 #include <hip/hip_complex.h>
@@ -162,10 +165,10 @@ size_t sharedMemoryBGijHip(LSMSSystemParameters &lsms, size_t *hfnOffset, size_t
   size += sizeof(double) * (2*lsms.maxlmax + 1);
 
   *plmOffset = size;
-  size += sizeof(double) * (lsms.angularMomentumIndices.ndlm);
+  size += sizeof(double) * (AngularMomentumIndices::ndlm);
 
   // *dlmOffset = size;
-  // size += sizeof(deviceDoubleComplex) * (lsms.angularMomentumIndices.ndlj);
+  // size += sizeof(deviceDoubleComplex) * (AngularMomentumIndices::ndlj);
   
   return size;
 }
@@ -251,9 +254,9 @@ void buildGijHipKernel(Real *LIZPos, int *LIZlmax, int *lofk, int *mofk, deviceD
     Real *sinmp = (Real *) (sharedMemory + sinmpOffset);
     // Real cosmp[2*lsms.maxlmax + 1];
     Real *cosmp = (Real *) (sharedMemory + cosmpOffset);
-    // Real plm[lsms.angularMomentumIndices.ndlm];
+    // Real plm[AngularMomentumIndices::ndlm];
     Real *plm = (Real *) (sharedMemory + plmOffset);
-    // Complex dlm[lsms.angularMomentumIndices.ndlj];
+    // Complex dlm[AngularMomentumIndices::ndlj];
     // deviceDoubleComplex *dlm = (deviceDoubleComplex *) (sharedMemory + dlmOffset);
 
 #if defined(COMPARE_ORIGINAL)
@@ -335,7 +338,12 @@ void buildGijHipKernel(Real *LIZPos, int *LIZlmax, int *lofk, int *mofk, deviceD
     {
       int lm2 = ij % kkri;
       int lm1 = ij / kkri;
-      devBgij[IDX(iOffset + lm2, jOffset + lm1, nrmat_ns)] = make_hipDoubleComplex(0.0, 0.0);
+
+      // auto devBgijPointer = &devBgij[IDX(iOffset + lm2, jOffset + lm1, nrmat_ns)];
+      
+      // devBgij[IDX(iOffset + lm2, jOffset + lm1, nrmat_ns)] = make_hipDoubleComplex(0.0, 0.0);
+      // *devBgijPointer =  make_hipDoubleComplex(0.0, 0.0);
+      auto devBgijValue =  make_hipDoubleComplex(0.0, 0.0);
       // bgij(iOffset + lm2, jOffset + lm1) = 0.0;
       // }
     
@@ -361,16 +369,24 @@ void buildGijHipKernel(Real *LIZPos, int *LIZlmax, int *lofk, int *mofk, deviceD
       int m3=m2-m1;
       int llow=max(abs(m3), abs(l1-l2));
       if(hipCabs(prel)==0.0) llow=l1+l2;
+
+      int cgntIdx0 = IDX3(0,lm1,lm2,lmaxp1_cgnt,ndlj_cgnt);
+      
       for(int l3=l1+l2; l3>=llow; l3-=2)
       {
         int j=l3*(l3+1)+m3;
         // gij[lm2+lm1*kkri] = gij[lm2+lm1*kkri]+cgnt(l3/2,lm1,lm2)*dlm[j];
-        devBgij[IDX(iOffset + lm2, jOffset + lm1, nrmat_ns)] =  devBgij[IDX(iOffset + lm2, jOffset + lm1, nrmat_ns)]
-          + cgnt[IDX3(l3/2,lm1,lm2,lmaxp1_cgnt,ndlj_cgnt)]
+        // devBgij[IDX(iOffset + lm2, jOffset + lm1, nrmat_ns)] =  devBgij[IDX(iOffset + lm2, jOffset + lm1, nrmat_ns)]
+        // *devBgijPointer = *devBgijPointer
+        devBgijValue = devBgijValue
+          // + cgnt[IDX3(l3/2,lm1,lm2,lmaxp1_cgnt,ndlj_cgnt)]
+          + cgnt[(l3 >> 1) + cgntIdx0]
           * dlmFunction(hfn, cosmp, sinmp, plm, l3, m3); //dlm[j];
       }
       // gij[lm2+lm1*kkri]=pi4*illp(lm2,lm1)*gij[lm2+lm1*kkri];
-      devBgij[IDX(iOffset + lm2, jOffset + lm1, nrmat_ns)] = devBgij[IDX(iOffset + lm2, jOffset + lm1, nrmat_ns)]
+      // devBgij[IDX(iOffset + lm2, jOffset + lm1, nrmat_ns)] = devBgij[IDX(iOffset + lm2, jOffset + lm1, nrmat_ns)]
+      // *devBgijPointer = *devBgijPointer
+      devBgij[IDX(iOffset + lm2, jOffset + lm1, nrmat_ns)] = devBgijValue
         * pi4 * illp[IDX(lm2, lm1, ndlj_illp)];
     }
 
@@ -403,9 +419,11 @@ void buildTmatNHip(int ispin, int n_spin_pola, int n_spin_cant, int iie, int blk
             int jm=jsm+kkrsz_ns*j+kkrsz*is;
 //                int one=1;
 //                BLAS::zcopy_(&kkr1,&local.tmatStore(iie*local.blkSizeTmatStore+jm,atom.LIZStoreIdx[ir1]),&one,&tmat_n[im],&one);
+            int tmatIdx0 = IDX(iie*blkSizeTmatStore + jm, lizStoreIdx, tmatStoreLDim);
             for(int i=0; i<kkr1; i++)
             {
-              tmat_n[im+i] = devTmatStore[IDX(iie*blkSizeTmatStore+jm+i, lizStoreIdx, tmatStoreLDim)];
+              // tmat_n[im+i] = devTmatStore[IDX(iie*blkSizeTmatStore+jm+i, lizStoreIdx, tmatStoreLDim)];
+              tmat_n[im+i] = devTmatStore[i + tmatIdx0];
             }
             im+=kkr1;
           }
@@ -422,9 +440,11 @@ void buildTmatNHip(int ispin, int n_spin_pola, int n_spin_cant, int iie, int blk
         int jm=jsm+kkrsz_ns*j;
 //           int one=1;
 //           BLAS::zcopy_(&kkr1,&local.tmatStore(iie*local.blkSizeTmatStore+jm,atom.LIZStoreIdx[ir1]),&one,&tmat_n[im],&one);
+        int tmatIdx0 = IDX(iie*blkSizeTmatStore + jm, lizStoreIdx, tmatStoreLDim);
         for(int i=0; i<kkr1; i++)
         {
-          tmat_n[im+i] = devTmatStore[IDX(iie*blkSizeTmatStore+jm+i, lizStoreIdx, tmatStoreLDim)];
+          // tmat_n[im+i] = devTmatStore[IDX(iie*blkSizeTmatStore  +jm + i, lizStoreIdx, tmatStoreLDim)];
+          tmat_n[im+i] = devTmatStore[i + tmatIdx0];
         }
         im+=kkr1;
       }
@@ -461,9 +481,11 @@ void buildKKRMatrixMultiplyKernelHip(int *LIZlmax, int *LIZStoreIdx, int *offset
 //                     // &bgijSmall(0, 0), &kkrsz_ns, &czero,
 //                     &m(iOffset, jOffset), &nrmat_ns);
 
+    /*
     for(int j=0; j<kkr2_ns; j++)
     {
     }
+    */
         
 //    for(int i=0; i<kkr1_ns; i++)
 //      for(int j=0; j<kkr2_ns; j++)
@@ -477,11 +499,24 @@ void buildKKRMatrixMultiplyKernelHip(int *LIZlmax, int *LIZStoreIdx, int *offset
       int i = ij % kkr1_ns;
       int j = ij / kkr1_ns;
 
-      devM[IDX(iOffset + i, jOffset + j, nrmat_ns)] = make_hipDoubleComplex(0.0,0.0);
+      // int ijIdx = IDX(iOffset + i, jOffset + j, nrmat_ns);
+      // auto devMPointer = &devM[ijIdx];
+      // devM[IDX(iOffset + i, jOffset + j, nrmat_ns)] = make_hipDoubleComplex(0.0,0.0);
+      // *devMPointer = make_hipDoubleComplex(0.0,0.0);
+      auto devMValue = make_hipDoubleComplex(0.0,0.0);
+      int devBgijIdx0 = IDX(iOffset, jOffset + j, nrmat_ns);
+      int tmat_nIdx = IDX(i,0,kkr1_ns);
       for(int k=0; k<kkr1_ns ; k++)
-        devM[IDX(iOffset + i, jOffset + j, nrmat_ns)] = devM[IDX(iOffset + i, jOffset + j, nrmat_ns)] -
-          tmat_n[IDX(i,k,kkr1_ns)] * // tmat_n(i, k) * // local.tmatStore(iie*local.blkSizeTmatStore + , atom.LIZStoreIdx[ir1]) *
-          devBgij[IDX(iOffset + k, jOffset + j, nrmat_ns)];
+      {
+        // devM[IDX(iOffset + i, jOffset + j, nrmat_ns)] = devM[IDX(iOffset + i, jOffset + j, nrmat_ns)] -
+        devMValue = devMValue -
+          // tmat_n[IDX(i,k,kkr1_ns)] *
+          tmat_n[tmat_nIdx] *
+          // devBgij[IDX(iOffset + k, jOffset + j, nrmat_ns)];
+          devBgij[k + devBgijIdx0];
+        tmat_nIdx += kkr1_ns;
+      }
+      devM[IDX(iOffset + i, jOffset + j, nrmat_ns)] = devMValue;
     }
     
   }
@@ -561,13 +596,13 @@ void buildKKRMatrixLMaxIdenticalHip(LSMSSystemParameters &lsms, LocalTypeInfo &l
   Real testSinmp[2*lsms.maxlmax + 1];
   Real testCosmp[2*lsms.maxlmax + 1];
   // Real plm[((lsms.maxlmax+1) * (lsms.maxlmax+2)) / 2];
-  Real testPlm[lsms.angularMomentumIndices.ndlm];
-// Complex testDlm[lsms.angularMomentumIndices.ndlj];
+  Real testPlm[AngularMomentumIndices::ndlm];
+// Complex testDlm[AngularMomentumIndices::ndlj];
   deviceMemcpy(testHfn, devTestSM + hfnOffset, (2*lsms.maxlmax + 1)*sizeof(Complex), deviceMemcpyDeviceToHost);
   deviceMemcpy(testSinmp, devTestSM + sinmpOffset, (2*lsms.maxlmax + 1)*sizeof(Real), deviceMemcpyDeviceToHost);
   deviceMemcpy(testCosmp, devTestSM + cosmpOffset, (2*lsms.maxlmax + 1)*sizeof(Real), deviceMemcpyDeviceToHost);
-  deviceMemcpy(testPlm, devTestSM + plmOffset, lsms.angularMomentumIndices.ndlm*sizeof(Real), deviceMemcpyDeviceToHost);
-// deviceMemcpy(testDlm, devTestSM + dlmOffset, lsms.angularMomentumIndices.ndlj*sizeof(Complex), deviceMemcpyDeviceToHost);
+  deviceMemcpy(testPlm, devTestSM + plmOffset, AngularMomentumIndices::ndlm*sizeof(Real), deviceMemcpyDeviceToHost);
+// deviceMemcpy(testDlm, devTestSM + dlmOffset, AngularMomentumIndices::ndlj*sizeof(Complex), deviceMemcpyDeviceToHost);
 
   for(int i = 0; i < atom.numLIZ; i++)
   {
@@ -585,8 +620,8 @@ void buildKKRMatrixLMaxIdenticalHip(LSMSSystemParameters &lsms, LocalTypeInfo &l
   Real sinmp[2*lsms.maxlmax + 1];
   Real cosmp[2*lsms.maxlmax + 1];
   // Real plm[((lsms.maxlmax+1) * (lsms.maxlmax+2)) / 2];
-  Real plm[lsms.angularMomentumIndices.ndlm];
-  Complex dlm[lsms.angularMomentumIndices.ndlj];
+  Real plm[AngularMomentumIndices::ndlm];
+  Complex dlm[AngularMomentumIndices::ndlj];
   Real rij[3];
   Real pi4=4.0*2.0*std::asin(1.0);
   bool exitCompare = false;
@@ -613,11 +648,11 @@ void buildKKRMatrixLMaxIdenticalHip(LSMSSystemParameters &lsms, LocalTypeInfo &l
         int lmax=lsms.maxlmax;
         int kkrsz=(lmax+1)*(lmax+1);
         makegij_(&atom.LIZlmax[ir1],&kkr1,&atom.LIZlmax[ir2],&kkr2,
-                 &lsms.maxlmax,&kkrsz,&lsms.angularMomentumIndices.ndlj,&lsms.angularMomentumIndices.ndlm,
+                 &lsms.maxlmax,&kkrsz,&AngularMomentumIndices::ndlj,&AngularMomentumIndices::ndlm,
                  &prel,&rij[0],&sinmp[0],&cosmp[0],
                  &sphericalHarmonicsCoeficients.clm[0],&plm[0],
                  &gauntCoeficients.cgnt(0,0,0),&gauntCoeficients.lmax,
-                 &lsms.angularMomentumIndices.lofk[0],&lsms.angularMomentumIndices.mofk[0],
+                 &AngularMomentumIndices::lofk[0],&AngularMomentumIndices::mofk[0],
                  &iFactors.ilp1[0],&iFactors.illp(0,0),
                  &hfn[0],&dlm[0],&gijTest(0,0),
                  &pi4,&lsms.global.iprint,lsms.global.istop,32);
